Change code style rules and Refactor dev script
