+++ .prettierrc
@@ -1,0 +1,6 @@
+{
+  "arrowParens": "always",
+  "printWidth": 120,
+  "singleQuote": true,
+  "trailingComma": "es5"
+}
