+++ package.json
@@ -9,10 +9,9 @@
     "build:front": "parcel build --target electron --public-url ./ index.html",
     "build": "npm run build:front && npm run build:back",
     "lint": "tslint src/index.tsx",
-    "serve": "parcel index.html",
     "test": "jest",
     "test:watch": "jest --watch",
-    "dev": "node ./dev.js"
+    "dev": "node ./launchDevEnvironment.js"
   },
   "repository": {
     "type": "git",
@@ -30,7 +29,6 @@
     "@types/mock-fs": "^3.6.30",
     "@types/react": "^16.8.13",
     "@types/react-dom": "^16.8.4",
-    "electron": "^4.1.4",
     "jest": "^24.7.1",
     "mock-fs": "^4.8.0",
     "parcel-bundler": "^1.12.3",
@@ -43,6 +41,7 @@
   "dependencies": {
     "@material-ui/core": "^3.9.3",
     "react": "^16.8.6",
+    "electron": "^4.1.4",
     "react-dom": "^16.8.6"
   }
 }
