+++ tslint.json
@@ -1,20 +1,14 @@
 {
   "defaultSeverity": "error",
-  "extends": ["tslint:recommended", "tslint-config-prettier"],
+  "extends": ["tslint:latest", "tslint-config-prettier"],
   "jsRules": {
     "no-unused-expression": true,
-    "quotemark": [true, "single"],
-    "ordered-imports": [false],
-    "max-line-length": [true, 150],
     "member-ordering": [false],
     "interface-name": [false],
     "object-literal-sort-keys": false,
     "arrow-parens": true
   },
   "rules": {
-    "quotemark": [true, "single", "jsx-double"],
-    "ordered-imports": [false],
-    "max-line-length": [true, 150],
     "member-ordering": [false],
     "interface-name": [false],
     "object-literal-sort-keys": false,
